change to use preread

+++ internal/chunk/message.go
@@ -1,10 +1,10 @@
 package chunk
 
 type Message struct {
-	idx int
+	idx int64
 	msg []byte
 }
 
-func NewMessage(idx int, msg []byte) Message {
+func NewMessage(idx int64, msg []byte) Message {
 	return Message{idx: idx, msg: msg}
 }
